+++ utils.py
@@ -1,5 +1,7 @@
 """kytos/flow_manager utils."""
 
+from collections.abc import Callable
+
 from pyof.foundation.base import UBIntBase
 from pyof.v0x04.controller2switch.flow_mod import FlowModCommand
 
@@ -187,16 +189,16 @@ def validate_cookies_del(flows: list[dict]) -> None:
             )
 
 
-def flows_to_log_info(message: str, flow_dict: dict[str, list]) -> None:
+def flows_to_log(logger_fun: Callable, message: str, flow_dict: dict[str, list]):
     """Log flows, maximun flows in a log is 200"""
     length_msg = f"total_length: {len(flow_dict['flows'])}, "
     maximun = 200
     flows_n = len(flow_dict["flows"])
     i, j = 0, maximun
     while flow_dict["flows"][i:j]:
-        log.info(
+        logger_fun(
             f"{message}{length_msg} flows[{i}, {(j if j < flows_n else flows_n)}]:"
             f" {flow_dict['flows'][i:j]}"
         )
-        i, j = i + maximun, j + maximun
+        i, j = j, j + maximun
         length_msg = ""